WIP - moved pivotal wrapper to its own dir
